backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.1
langchain==0.1.6
langgraph==0.0.26
python-dotenv==1.0.0
pydantic==2.6.0
httpx==0.26.0
sendgrid==6.11.0
python-multipart==0.0.6
